+++ src/commonMain/kotlin/io/github/hansanto/kault/auth/oidc/common/OIDCResponseType.kt
@@ -1,0 +1,19 @@
+package io.github.hansanto.kault.auth.oidc.common
+
+import io.github.hansanto.kault.serializer.EnumSerializer
+import kotlinx.serialization.Serializable
+
+@Serializable(OIDCResponseTypeSerializer::class)
+public enum class OIDCResponseType(public val value: String) {
+    CODE("code"),
+    ID_TOKEN("id_token"),
+}
+
+/**
+ * Serializer for [OIDCResponseType].
+ */
+public object OIDCResponseTypeSerializer : EnumSerializer<OIDCResponseType>(
+    "kault.auth.oidc.common.OIDCResponseTypeSerializer",
+    OIDCResponseType.entries,
+    { it.value }
+)