+++ src/commonMain/kotlin/io/github/hansanto/kault/engine/kv/v2/payload/KvV2WriteRequest.kt
@@ -3,7 +3,6 @@ package io.github.hansanto.kault.engine.kv.v2.payload
 import io.github.hansanto.kault.BuilderDsl
 import io.github.hansanto.kault.KaultDsl
 import io.github.hansanto.kault.VaultClient
-import io.github.hansanto.kault.auth.approle.payload.AppRoleCreateCustomSecretIDPayload
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -39,7 +38,7 @@ public data class KvV2WriteRequest(
     )
 
     /**
-     * Builder class to simplify the creation of [AppRoleCreateCustomSecretIDPayload].
+     * Builder class to simplify the creation of [KvV2WriteRequest].
      */
     @KaultDsl
     public class Builder {